requirements.txt
python-telegram-bot==20.6
python-dotenv
cryptography
ccxt
requests